+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 
 - allow defaults to be a function on api
+- add content type header in fetch backend to match axios backend
 
 # 0.2.5